requirements.txt
torch
transformers
spacy
scikit-learn
numpy
tqdm
pandas
datasets
google-genai
python-dotenv
pytest
